requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
lxml==4.9.3
pandas==2.1.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
fake-useragent==1.4.0
requests-html==0.10.0
scrapy==2.11.0
aiohttp==3.9.1
asyncio==3.4.3
schedule==1.2.0
python-dateutil==2.8.2
regex==2023.10.3
urllib3==2.1.0
certifi==2023.11.17 